upd the back end interaction and models (still no good)

+++ lib/application/mappers/event_mapper.dart
@@ -22,7 +22,7 @@ abstract class EventMapper {
         description: model.description ?? '',
         location: model.location ?? '',
         datetime: model.occurringAt,
-        cost: model.cost.number,
+        cost: model.cost.number.toInt(),
         isOnline: model.onlineEvent,
         cover: model.coverUrl ?? '',
       );
